Rename operation.ts -> declaration.ts
